SE: Move InvocationCheck to Invocation
