+++ analyzers/src/SonarAnalyzer.Common/SymbolicExecution/Roslyn/RoslynSymbolicExecution.cs
@@ -54,7 +54,7 @@ namespace SonarAnalyzer.SymbolicExecution.Roslyn
             {
                 throw new ArgumentException("At least one check is expected", nameof(checks));
             }
-            this.checks = new(new SymbolicCheck[] { new ConstantCheck(), new InvocationCheck() }.Concat(checks).ToArray());
+            this.checks = new(new SymbolicCheck[] { new ConstantCheck() }.Concat(checks).ToArray());
             this.cancel = cancel;
             exceptionCandidate = new(cfg.OriginalOperation.ToSonar().SemanticModel.Compilation);
             lva = new(cfg, cancel);
